Add changes and total calcuation check, minor adjustments
