+++ Object Repository/API/backWebServices/VirtoCommerce.Order/OrderUpdate.rs
@@ -10,7 +10,7 @@
    <followRedirects>false</followRedirects>
    <httpBody></httpBody>
    <httpBodyContent>{
-  &quot;text&quot;: &quot;{\n  \&quot;customerId\&quot;: \&quot;1eb2fa8ac6574541afdb525833dadb46\&quot;,\n  \&quot;customerName\&quot;: \&quot;admin\&quot;,\n  \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n  \&quot;shoppingCartId\&quot;: \&quot;4cbaf9bf-834d-4306-bde6-af66af08054d\&quot;,\n  \&quot;isPrototype\&quot;: false,\n  \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.CustomerOrder\&quot;,\n  \&quot;addresses\&quot;: [\n    {\n      \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n      \&quot;key\&quot;: \&quot;fbc09e8e-d435-4a56-91d7-153a2beedc25\&quot;,\n      \&quot;name\&quot;: \&quot;qwe, ewr, 123123, Burundi\&quot;,\n      \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n      \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n      \&quot;city\&quot;: \&quot;ewr\&quot;,\n      \&quot;postalCode\&quot;: \&quot;123123\&quot;,\n      \&quot;line1\&quot;: \&quot;qwe\&quot;,\n      \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n      \&quot;lastName\&quot;: \&quot;qwe\&quot;,\n      \&quot;phone\&quot;: \&quot;+79817420529\&quot;,\n      \&quot;email\&quot;: \&quot;neateugene1@gmail.com\&quot;\n    }\n  ],\n  \&quot;inPayments\&quot;: [\n    {\n      \&quot;orderId\&quot;: \&quot;25333618-21cb-4518-919c-0c0e46d7a921\&quot;,\n      \&quot;gatewayCode\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n      \&quot;paymentMethod\&quot;: {\n        \&quot;name\&quot;: \&quot;Test payment method\&quot;,\n        \&quot;paymentMethodType\&quot;: \&quot;Unknown\&quot;,\n        \&quot;paymentMethodGroupType\&quot;: \&quot;Manual\&quot;,\n        \&quot;code\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 0,\n        \&quot;isAvailableForPartial\&quot;: false,\n        \&quot;price\&quot;: 0,\n        \&quot;priceWithTax\&quot;: 0,\n        \&quot;total\&quot;: 0,\n        \&quot;totalWithTax\&quot;: 0,\n        \&quot;discountAmount\&quot;: 0,\n        \&quot;discountAmountWithTax\&quot;: 0,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;typeName\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;,\n            \&quot;objectType\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Payment\&quot;,\n            \&quot;groupName\&quot;: \&quot;Payment|DefaultManualPaymentMethod\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Payment.DefaultManualPaymentMethod.ExampleSetting\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;ShortText\&quot;,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;taxTotal\&quot;: 0,\n        \&quot;taxPercentRate\&quot;: 0,\n        \&quot;id\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;\n      },\n      \&quot;customerId\&quot;: \&quot;1eb2fa8ac6574541afdb525833dadb46\&quot;,\n      \&quot;billingAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;cca9f7a6-d319-49ff-b3ba-1f75ba6c3678\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, ewr, 123123, Burundi\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;ewr\&quot;,\n        \&quot;postalCode\&quot;: \&quot;123123\&quot;,\n        \&quot;line1\&quot;: \&quot;qwe\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;qwe\&quot;,\n        \&quot;phone\&quot;: \&quot;+79817420529\&quot;,\n        \&quot;email\&quot;: \&quot;neateugene1@gmail.com\&quot;\n      },\n      \&quot;paymentStatus\&quot;: \&quot;New\&quot;,\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.PaymentIn\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;transactions\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;PaymentIn\&quot;,\n      \&quot;number\&quot;: \&quot;PI210803-00001\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 995.99,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-03T09:13:20.1080876Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-03T09:13:20.1080876Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;82a112c6-5440-4f92-ad22-4046f89444ab\&quot;\n    }\n  ],\n  \&quot;items\&quot;: [\n    {\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;price\&quot;: 995.99,\n      \&quot;priceWithTax\&quot;: 995.99,\n      \&quot;placedPrice\&quot;: 995.99,\n      \&quot;placedPriceWithTax\&quot;: 995.99,\n      \&quot;extendedPrice\&quot;: 995.99,\n      \&quot;extendedPriceWithTax\&quot;: 995.99,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;discountTotal\&quot;: 0,\n      \&quot;discountTotalWithTax\&quot;: 0,\n      \&quot;fee\&quot;: 0,\n      \&quot;feeWithTax\&quot;: 0,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;reserveQuantity\&quot;: 0,\n      \&quot;quantity\&quot;: \&quot;${quantity}\&quot;,\n      \&quot;productId\&quot;: \&quot;9cbd8f316e254a679ba34a900fccb076\&quot;,\n      \&quot;sku\&quot;: \&quot;3DRSOLO\&quot;,\n      \&quot;productType\&quot;: \&quot;Physical\&quot;,\n      \&quot;catalogId\&quot;: \&quot;4974648a41df4e6ea67ef2ad76d7bbd4\&quot;,\n      \&quot;categoryId\&quot;: \&quot;e51b5f9eea094a44939c11d4d4fa3bb1\&quot;,\n      \&quot;name\&quot;: \&quot;3DR Solo Quadcopter (No Gimbal)\&quot;,\n      \&quot;imageUrl\&quot;: \&quot;//localhost:9112/assets/catalog/3DRSOLO/1428965138000_1133723.jpg\&quot;,\n      \&quot;isGift\&quot;: false,\n      \&quot;isCancelled\&quot;: false,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.LineItem\&quot;,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;taxDetails\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-03T09:13:20.108155Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-03T09:13:20.108155Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;9e30a534-6f21-4b40-9341-e4d872138455\&quot;\n    }\n  ],\n  \&quot;shipments\&quot;: [\n    {\n      \&quot;shipmentMethodCode\&quot;: \&quot;FixedRate\&quot;,\n      \&quot;shipmentMethodOption\&quot;: \&quot;Ground\&quot;,\n      \&quot;shippingMethod\&quot;: {\n        \&quot;code\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;name\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 1,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Ground.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          },\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Air.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;typeName\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n        \&quot;id\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;\n      },\n      \&quot;customerOrderId\&quot;: \&quot;25333618-21cb-4518-919c-0c0e46d7a921\&quot;,\n      \&quot;items\&quot;: [],\n      \&quot;packages\&quot;: [],\n      \&quot;inPayments\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;deliveryAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;cb7dc188-ef13-4d8c-b2ca-43d6372ef2fd\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, ewr, 123123, Burundi\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;ewr\&quot;,\n        \&quot;postalCode\&quot;: \&quot;123123\&quot;,\n        \&quot;line1\&quot;: \&quot;qwe\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;qwe\&quot;,\n        \&quot;phone\&quot;: \&quot;+79817420529\&quot;,\n        \&quot;email\&quot;: \&quot;neateugene1@gmail.com\&quot;\n      },\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;fee\&quot;: 0,\n      \&quot;feeWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.Shipment\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;Shipment\&quot;,\n      \&quot;number\&quot;: \&quot;SH210803-00001\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 0,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-03T09:13:20.1081876Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-03T09:13:20.1081876Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;789fa160-022f-4f09-aae8-c1fbb2b2c661\&quot;\n    }\n  ],\n  \&quot;discounts\&quot;: [],\n  \&quot;discountAmount\&quot;: 0,\n  \&quot;taxDetails\&quot;: [],\n  \&quot;total\&quot;: 995.99,\n  \&quot;subTotal\&quot;: 995.99,\n  \&quot;subTotalWithTax\&quot;: 995.99,\n  \&quot;subTotalDiscount\&quot;: 0,\n  \&quot;subTotalDiscountWithTax\&quot;: 0,\n  \&quot;subTotalTaxTotal\&quot;: 0,\n  \&quot;shippingTotal\&quot;: 0,\n  \&quot;shippingTotalWithTax\&quot;: 0,\n  \&quot;shippingSubTotal\&quot;: 0,\n  \&quot;shippingSubTotalWithTax\&quot;: 0,\n  \&quot;shippingDiscountTotal\&quot;: 0,\n  \&quot;shippingDiscountTotalWithTax\&quot;: 0,\n  \&quot;shippingTaxTotal\&quot;: 0,\n  \&quot;paymentTotal\&quot;: 0,\n  \&quot;paymentTotalWithTax\&quot;: 0,\n  \&quot;paymentSubTotal\&quot;: 0,\n  \&quot;paymentSubTotalWithTax\&quot;: 0,\n  \&quot;paymentDiscountTotal\&quot;: 0,\n  \&quot;paymentDiscountTotalWithTax\&quot;: 0,\n  \&quot;paymentTaxTotal\&quot;: 0,\n  \&quot;discountTotal\&quot;: 0,\n  \&quot;discountTotalWithTax\&quot;: 0,\n  \&quot;fee\&quot;: 0,\n  \&quot;feeWithTax\&quot;: 0,\n  \&quot;feeTotal\&quot;: 0,\n  \&quot;feeTotalWithTax\&quot;: 0,\n  \&quot;taxTotal\&quot;: 0,\n  \&quot;taxPercentRate\&quot;: 0,\n  \&quot;languageCode\&quot;: \&quot;en-US\&quot;,\n  \&quot;operationType\&quot;: \&quot;CustomerOrder\&quot;,\n  \&quot;number\&quot;: \&quot;CO210803-00001\&quot;,\n  \&quot;isApproved\&quot;: false,\n  \&quot;status\&quot;: \&quot;Paid\&quot;,\n  \&quot;currency\&quot;: \&quot;USD\&quot;,\n  \&quot;sum\&quot;: 995.99,\n  \&quot;childrenOperations\&quot;: [\n    {\n      \&quot;orderId\&quot;: \&quot;25333618-21cb-4518-919c-0c0e46d7a921\&quot;,\n      \&quot;gatewayCode\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n      \&quot;paymentMethod\&quot;: {\n        \&quot;name\&quot;: \&quot;Test payment method\&quot;,\n        \&quot;paymentMethodType\&quot;: \&quot;Unknown\&quot;,\n        \&quot;paymentMethodGroupType\&quot;: \&quot;Manual\&quot;,\n        \&quot;code\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 0,\n        \&quot;isAvailableForPartial\&quot;: false,\n        \&quot;price\&quot;: 0,\n        \&quot;priceWithTax\&quot;: 0,\n        \&quot;total\&quot;: 0,\n        \&quot;totalWithTax\&quot;: 0,\n        \&quot;discountAmount\&quot;: 0,\n        \&quot;discountAmountWithTax\&quot;: 0,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;typeName\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;,\n            \&quot;objectType\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Payment\&quot;,\n            \&quot;groupName\&quot;: \&quot;Payment|DefaultManualPaymentMethod\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Payment.DefaultManualPaymentMethod.ExampleSetting\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;ShortText\&quot;,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;taxTotal\&quot;: 0,\n        \&quot;taxPercentRate\&quot;: 0,\n        \&quot;id\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;\n      },\n      \&quot;customerId\&quot;: \&quot;1eb2fa8ac6574541afdb525833dadb46\&quot;,\n      \&quot;billingAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;cca9f7a6-d319-49ff-b3ba-1f75ba6c3678\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, ewr, 123123, Burundi\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;ewr\&quot;,\n        \&quot;postalCode\&quot;: \&quot;123123\&quot;,\n        \&quot;line1\&quot;: \&quot;qwe\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;qwe\&quot;,\n        \&quot;phone\&quot;: \&quot;+79817420529\&quot;,\n        \&quot;email\&quot;: \&quot;neateugene1@gmail.com\&quot;\n      },\n      \&quot;paymentStatus\&quot;: \&quot;New\&quot;,\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.PaymentIn\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;transactions\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;PaymentIn\&quot;,\n      \&quot;number\&quot;: \&quot;PI210803-00001\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 995.99,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-03T09:13:20.1080876Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-03T09:13:20.1080876Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;82a112c6-5440-4f92-ad22-4046f89444ab\&quot;\n    },\n    {\n      \&quot;shipmentMethodCode\&quot;: \&quot;FixedRate\&quot;,\n      \&quot;shipmentMethodOption\&quot;: \&quot;Ground\&quot;,\n      \&quot;shippingMethod\&quot;: {\n        \&quot;code\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;name\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 1,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Ground.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          },\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Air.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;typeName\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n        \&quot;id\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;\n      },\n      \&quot;customerOrderId\&quot;: \&quot;25333618-21cb-4518-919c-0c0e46d7a921\&quot;,\n      \&quot;items\&quot;: [],\n      \&quot;packages\&quot;: [],\n      \&quot;inPayments\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;deliveryAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;cb7dc188-ef13-4d8c-b2ca-43d6372ef2fd\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, ewr, 123123, Burundi\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;ewr\&quot;,\n        \&quot;postalCode\&quot;: \&quot;123123\&quot;,\n        \&quot;line1\&quot;: \&quot;qwe\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;qwe\&quot;,\n        \&quot;phone\&quot;: \&quot;+79817420529\&quot;,\n        \&quot;email\&quot;: \&quot;neateugene1@gmail.com\&quot;\n      },\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;fee\&quot;: 0,\n      \&quot;feeWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.Shipment\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;Shipment\&quot;,\n      \&quot;number\&quot;: \&quot;SH210803-00001\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 0,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-03T09:13:20.1081876Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-03T09:13:20.1081876Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;789fa160-022f-4f09-aae8-c1fbb2b2c661\&quot;\n    }\n  ],\n  \&quot;isCancelled\&quot;: false,\n  \&quot;dynamicProperties\&quot;: [],\n  \&quot;createdDate\&quot;: \&quot;2021-08-03T09:13:20.1079407Z\&quot;,\n  \&quot;modifiedDate\&quot;: \&quot;2021-08-03T09:13:20.2683361Z\&quot;,\n  \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n  \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n  \&quot;id\&quot;: \&quot;${orderId}\&quot;\n}&quot;,
+  &quot;text&quot;: &quot;{\n  \&quot;customerId\&quot;: \&quot;8f20288b-da0a-45d7-8ade-bce2adbadcfe\&quot;,\n  \&quot;customerName\&quot;: \&quot;${userName}\&quot;,\n  \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n  \&quot;shoppingCartId\&quot;: \&quot;44f53228-1251-4cb7-a284-58e37a1be948\&quot;,\n  \&quot;isPrototype\&quot;: false,\n  \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.CustomerOrder\&quot;,\n  \&quot;addresses\&quot;: [\n    {\n      \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n      \&quot;key\&quot;: \&quot;bbb0b571-c387-4541-a9ae-aba42993550b\&quot;,\n      \&quot;name\&quot;: \&quot;qwe, 123, qwerty, qwe, 90201, Burundi\&quot;,\n      \&quot;organization\&quot;: \&quot;qwe\&quot;,\n      \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n      \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n      \&quot;city\&quot;: \&quot;qwe\&quot;,\n      \&quot;postalCode\&quot;: \&quot;90201\&quot;,\n      \&quot;line1\&quot;: \&quot;qwerty\&quot;,\n      \&quot;line2\&quot;: \&quot;123\&quot;,\n      \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n      \&quot;lastName\&quot;: \&quot;rty\&quot;,\n      \&quot;phone\&quot;: \&quot;+79991234567\&quot;,\n      \&quot;email\&quot;: \&quot;qwerty@mail.com\&quot;\n    }\n  ],\n  \&quot;inPayments\&quot;: [\n    {\n      \&quot;orderId\&quot;: \&quot;8383028c-4d80-46bc-887e-282d0707a070\&quot;,\n      \&quot;gatewayCode\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n      \&quot;paymentMethod\&quot;: {\n        \&quot;name\&quot;: \&quot;Test payment method\&quot;,\n        \&quot;paymentMethodType\&quot;: \&quot;Unknown\&quot;,\n        \&quot;paymentMethodGroupType\&quot;: \&quot;Manual\&quot;,\n        \&quot;code\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 0,\n        \&quot;isAvailableForPartial\&quot;: false,\n        \&quot;price\&quot;: 0,\n        \&quot;priceWithTax\&quot;: 0,\n        \&quot;total\&quot;: 0,\n        \&quot;totalWithTax\&quot;: 0,\n        \&quot;discountAmount\&quot;: 0,\n        \&quot;discountAmountWithTax\&quot;: 0,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;typeName\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;,\n            \&quot;objectType\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Payment\&quot;,\n            \&quot;groupName\&quot;: \&quot;Payment|DefaultManualPaymentMethod\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Payment.DefaultManualPaymentMethod.ExampleSetting\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;ShortText\&quot;,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;taxTotal\&quot;: 0,\n        \&quot;taxPercentRate\&quot;: 0,\n        \&quot;id\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;\n      },\n      \&quot;customerId\&quot;: \&quot;8f20288b-da0a-45d7-8ade-bce2adbadcfe\&quot;,\n      \&quot;billingAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;8ddd9005-2d75-4978-8ba3-9be99e5109a7\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, 123, qwerty, qwe, 90201, Burundi\&quot;,\n        \&quot;organization\&quot;: \&quot;qwe\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;qwe\&quot;,\n        \&quot;postalCode\&quot;: \&quot;90201\&quot;,\n        \&quot;line1\&quot;: \&quot;qwerty\&quot;,\n        \&quot;line2\&quot;: \&quot;123\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;rty\&quot;,\n        \&quot;phone\&quot;: \&quot;+79991234567\&quot;,\n        \&quot;email\&quot;: \&quot;qwerty@mail.com\&quot;\n      },\n      \&quot;paymentStatus\&quot;: \&quot;New\&quot;,\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.PaymentIn\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;transactions\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;PaymentIn\&quot;,\n      \&quot;number\&quot;: \&quot;PI210804-00004\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 995.99,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-04T12:27:58.603881Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-04T12:27:58.603881Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;c3c7265e-9a26-4760-af66-f4e3c0d0f1dc\&quot;\n    }\n  ],\n  \&quot;items\&quot;: [\n    {\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;price\&quot;: 995.99,\n      \&quot;priceWithTax\&quot;: 995.99,\n      \&quot;placedPrice\&quot;: 995.99,\n      \&quot;placedPriceWithTax\&quot;: 995.99,\n      \&quot;extendedPrice\&quot;: 995.99,\n      \&quot;extendedPriceWithTax\&quot;: 995.99,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;discountTotal\&quot;: 0,\n      \&quot;discountTotalWithTax\&quot;: 0,\n      \&quot;fee\&quot;: 0,\n      \&quot;feeWithTax\&quot;: 0,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;reserveQuantity\&quot;: 0,\n      \&quot;quantity\&quot;: \&quot;${quantity}\&quot;,\n      \&quot;productId\&quot;: \&quot;9cbd8f316e254a679ba34a900fccb076\&quot;,\n      \&quot;sku\&quot;: \&quot;3DRSOLO\&quot;,\n      \&quot;productType\&quot;: \&quot;Physical\&quot;,\n      \&quot;catalogId\&quot;: \&quot;4974648a41df4e6ea67ef2ad76d7bbd4\&quot;,\n      \&quot;categoryId\&quot;: \&quot;e51b5f9eea094a44939c11d4d4fa3bb1\&quot;,\n      \&quot;name\&quot;: \&quot;3DR Solo Quadcopter (No Gimbal)\&quot;,\n      \&quot;imageUrl\&quot;: \&quot;//localhost:9112/assets/catalog/3DRSOLO/1428965138000_1133723.jpg\&quot;,\n      \&quot;isGift\&quot;: false,\n      \&quot;isCancelled\&quot;: false,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.LineItem\&quot;,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;taxDetails\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-04T12:27:58.6039371Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-04T12:27:58.6039371Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;e6896b30-b538-4961-a0d7-1cbfcf4e5494\&quot;\n    }\n  ],\n  \&quot;shipments\&quot;: [\n    {\n      \&quot;shipmentMethodCode\&quot;: \&quot;FixedRate\&quot;,\n      \&quot;shipmentMethodOption\&quot;: \&quot;Ground\&quot;,\n      \&quot;shippingMethod\&quot;: {\n        \&quot;code\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;name\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 1,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Ground.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          },\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Air.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;typeName\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n        \&quot;id\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;\n      },\n      \&quot;customerOrderId\&quot;: \&quot;8383028c-4d80-46bc-887e-282d0707a070\&quot;,\n      \&quot;items\&quot;: [],\n      \&quot;packages\&quot;: [],\n      \&quot;inPayments\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;deliveryAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;191f38d3-bb23-4599-9c06-88ce18323e76\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, 123, qwerty, qwe, 90201, Burundi\&quot;,\n        \&quot;organization\&quot;: \&quot;qwe\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;qwe\&quot;,\n        \&quot;postalCode\&quot;: \&quot;90201\&quot;,\n        \&quot;line1\&quot;: \&quot;qwerty\&quot;,\n        \&quot;line2\&quot;: \&quot;123\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;rty\&quot;,\n        \&quot;phone\&quot;: \&quot;+79991234567\&quot;,\n        \&quot;email\&quot;: \&quot;qwerty@mail.com\&quot;\n      },\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;fee\&quot;: 0,\n      \&quot;feeWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.Shipment\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;Shipment\&quot;,\n      \&quot;number\&quot;: \&quot;SH210804-00003\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 0,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-04T12:27:58.6040113Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-04T12:27:58.6040113Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;61b8788d-cd41-4320-914c-27ea1308dee8\&quot;\n    }\n  ],\n  \&quot;discounts\&quot;: [],\n  \&quot;discountAmount\&quot;: 0,\n  \&quot;taxDetails\&quot;: [],\n  \&quot;total\&quot;: 995.99,\n  \&quot;subTotal\&quot;: 995.99,\n  \&quot;subTotalWithTax\&quot;: 995.99,\n  \&quot;subTotalDiscount\&quot;: 0,\n  \&quot;subTotalDiscountWithTax\&quot;: 0,\n  \&quot;subTotalTaxTotal\&quot;: 0,\n  \&quot;shippingTotal\&quot;: 0,\n  \&quot;shippingTotalWithTax\&quot;: 0,\n  \&quot;shippingSubTotal\&quot;: 0,\n  \&quot;shippingSubTotalWithTax\&quot;: 0,\n  \&quot;shippingDiscountTotal\&quot;: 0,\n  \&quot;shippingDiscountTotalWithTax\&quot;: 0,\n  \&quot;shippingTaxTotal\&quot;: 0,\n  \&quot;paymentTotal\&quot;: 0,\n  \&quot;paymentTotalWithTax\&quot;: 0,\n  \&quot;paymentSubTotal\&quot;: 0,\n  \&quot;paymentSubTotalWithTax\&quot;: 0,\n  \&quot;paymentDiscountTotal\&quot;: 0,\n  \&quot;paymentDiscountTotalWithTax\&quot;: 0,\n  \&quot;paymentTaxTotal\&quot;: 0,\n  \&quot;discountTotal\&quot;: 0,\n  \&quot;discountTotalWithTax\&quot;: 0,\n  \&quot;fee\&quot;: 0,\n  \&quot;feeWithTax\&quot;: 0,\n  \&quot;feeTotal\&quot;: 0,\n  \&quot;feeTotalWithTax\&quot;: 0,\n  \&quot;taxTotal\&quot;: 0,\n  \&quot;taxPercentRate\&quot;: 0,\n  \&quot;languageCode\&quot;: \&quot;en-US\&quot;,\n  \&quot;operationType\&quot;: \&quot;CustomerOrder\&quot;,\n  \&quot;number\&quot;: \&quot;CO210804-00003\&quot;,\n  \&quot;isApproved\&quot;: false,\n  \&quot;status\&quot;: \&quot;Paid\&quot;,\n  \&quot;currency\&quot;: \&quot;USD\&quot;,\n  \&quot;sum\&quot;: 995.99,\n  \&quot;childrenOperations\&quot;: [\n    {\n      \&quot;orderId\&quot;: \&quot;8383028c-4d80-46bc-887e-282d0707a070\&quot;,\n      \&quot;gatewayCode\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n      \&quot;paymentMethod\&quot;: {\n        \&quot;name\&quot;: \&quot;Test payment method\&quot;,\n        \&quot;paymentMethodType\&quot;: \&quot;Unknown\&quot;,\n        \&quot;paymentMethodGroupType\&quot;: \&quot;Manual\&quot;,\n        \&quot;code\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 0,\n        \&quot;isAvailableForPartial\&quot;: false,\n        \&quot;price\&quot;: 0,\n        \&quot;priceWithTax\&quot;: 0,\n        \&quot;total\&quot;: 0,\n        \&quot;totalWithTax\&quot;: 0,\n        \&quot;discountAmount\&quot;: 0,\n        \&quot;discountAmountWithTax\&quot;: 0,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;typeName\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;,\n            \&quot;objectType\&quot;: \&quot;DefaultManualPaymentMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Payment\&quot;,\n            \&quot;groupName\&quot;: \&quot;Payment|DefaultManualPaymentMethod\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Payment.DefaultManualPaymentMethod.ExampleSetting\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;ShortText\&quot;,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;taxTotal\&quot;: 0,\n        \&quot;taxPercentRate\&quot;: 0,\n        \&quot;id\&quot;: \&quot;b53aff2cf71c43a1a4d69118f3f231c4\&quot;\n      },\n      \&quot;customerId\&quot;: \&quot;8f20288b-da0a-45d7-8ade-bce2adbadcfe\&quot;,\n      \&quot;billingAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;8ddd9005-2d75-4978-8ba3-9be99e5109a7\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, 123, qwerty, qwe, 90201, Burundi\&quot;,\n        \&quot;organization\&quot;: \&quot;qwe\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;qwe\&quot;,\n        \&quot;postalCode\&quot;: \&quot;90201\&quot;,\n        \&quot;line1\&quot;: \&quot;qwerty\&quot;,\n        \&quot;line2\&quot;: \&quot;123\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;rty\&quot;,\n        \&quot;phone\&quot;: \&quot;+79991234567\&quot;,\n        \&quot;email\&quot;: \&quot;qwerty@mail.com\&quot;\n      },\n      \&quot;paymentStatus\&quot;: \&quot;New\&quot;,\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.PaymentIn\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;transactions\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;PaymentIn\&quot;,\n      \&quot;number\&quot;: \&quot;PI210804-00004\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 995.99,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-04T12:27:58.603881Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-04T12:27:58.603881Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;c3c7265e-9a26-4760-af66-f4e3c0d0f1dc\&quot;\n    },\n    {\n      \&quot;shipmentMethodCode\&quot;: \&quot;FixedRate\&quot;,\n      \&quot;shipmentMethodOption\&quot;: \&quot;Ground\&quot;,\n      \&quot;shippingMethod\&quot;: {\n        \&quot;code\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;name\&quot;: \&quot;FixedRate\&quot;,\n        \&quot;logoUrl\&quot;: \&quot;https://raw.githubusercontent.com/VirtoCommerce/vc-module-core/master/VirtoCommerce.CoreModule.Web/Content/logoVC.png\&quot;,\n        \&quot;isActive\&quot;: true,\n        \&quot;priority\&quot;: 1,\n        \&quot;storeId\&quot;: \&quot;Electronics\&quot;,\n        \&quot;settings\&quot;: [\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Ground.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          },\n          {\n            \&quot;itHasValues\&quot;: false,\n            \&quot;objectId\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;,\n            \&quot;objectType\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n            \&quot;restartRequired\&quot;: false,\n            \&quot;moduleId\&quot;: \&quot;VirtoCommerce.Shipping\&quot;,\n            \&quot;groupName\&quot;: \&quot;Shipping|General\&quot;,\n            \&quot;name\&quot;: \&quot;VirtoCommerce.Shipping.FixedRateShippingMethod.Air.Rate\&quot;,\n            \&quot;isHidden\&quot;: false,\n            \&quot;valueType\&quot;: \&quot;Decimal\&quot;,\n            \&quot;defaultValue\&quot;: 0,\n            \&quot;isDictionary\&quot;: false\n          }\n        ],\n        \&quot;typeName\&quot;: \&quot;FixedRateShippingMethod\&quot;,\n        \&quot;id\&quot;: \&quot;fa8a1d7fa0ed4aec9358ef82596bb717\&quot;\n      },\n      \&quot;customerOrderId\&quot;: \&quot;8383028c-4d80-46bc-887e-282d0707a070\&quot;,\n      \&quot;items\&quot;: [],\n      \&quot;packages\&quot;: [],\n      \&quot;inPayments\&quot;: [],\n      \&quot;discounts\&quot;: [],\n      \&quot;deliveryAddress\&quot;: {\n        \&quot;addressType\&quot;: \&quot;BillingAndShipping\&quot;,\n        \&quot;key\&quot;: \&quot;191f38d3-bb23-4599-9c06-88ce18323e76\&quot;,\n        \&quot;name\&quot;: \&quot;qwe, 123, qwerty, qwe, 90201, Burundi\&quot;,\n        \&quot;organization\&quot;: \&quot;qwe\&quot;,\n        \&quot;countryCode\&quot;: \&quot;BDI\&quot;,\n        \&quot;countryName\&quot;: \&quot;Burundi\&quot;,\n        \&quot;city\&quot;: \&quot;qwe\&quot;,\n        \&quot;postalCode\&quot;: \&quot;90201\&quot;,\n        \&quot;line1\&quot;: \&quot;qwerty\&quot;,\n        \&quot;line2\&quot;: \&quot;123\&quot;,\n        \&quot;firstName\&quot;: \&quot;qwe\&quot;,\n        \&quot;lastName\&quot;: \&quot;rty\&quot;,\n        \&quot;phone\&quot;: \&quot;+79991234567\&quot;,\n        \&quot;email\&quot;: \&quot;qwerty@mail.com\&quot;\n      },\n      \&quot;price\&quot;: 0,\n      \&quot;priceWithTax\&quot;: 0,\n      \&quot;total\&quot;: 0,\n      \&quot;totalWithTax\&quot;: 0,\n      \&quot;discountAmount\&quot;: 0,\n      \&quot;discountAmountWithTax\&quot;: 0,\n      \&quot;fee\&quot;: 0,\n      \&quot;feeWithTax\&quot;: 0,\n      \&quot;objectType\&quot;: \&quot;VirtoCommerce.OrdersModule.Core.Model.Shipment\&quot;,\n      \&quot;taxTotal\&quot;: 0,\n      \&quot;taxPercentRate\&quot;: 0,\n      \&quot;taxDetails\&quot;: [],\n      \&quot;operationType\&quot;: \&quot;Shipment\&quot;,\n      \&quot;number\&quot;: \&quot;SH210804-00003\&quot;,\n      \&quot;isApproved\&quot;: false,\n      \&quot;status\&quot;: \&quot;New\&quot;,\n      \&quot;currency\&quot;: \&quot;USD\&quot;,\n      \&quot;sum\&quot;: 0,\n      \&quot;childrenOperations\&quot;: [],\n      \&quot;isCancelled\&quot;: false,\n      \&quot;dynamicProperties\&quot;: [],\n      \&quot;createdDate\&quot;: \&quot;2021-08-04T12:27:58.6040113Z\&quot;,\n      \&quot;modifiedDate\&quot;: \&quot;2021-08-04T12:27:58.6040113Z\&quot;,\n      \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n      \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n      \&quot;id\&quot;: \&quot;61b8788d-cd41-4320-914c-27ea1308dee8\&quot;\n    }\n  ],\n  \&quot;isCancelled\&quot;: false,\n  \&quot;dynamicProperties\&quot;: [],\n  \&quot;createdDate\&quot;: \&quot;2021-08-04T12:27:58.6037825Z\&quot;,\n  \&quot;modifiedDate\&quot;: \&quot;2021-08-04T12:27:58.7233157Z\&quot;,\n  \&quot;createdBy\&quot;: \&quot;admin\&quot;,\n  \&quot;modifiedBy\&quot;: \&quot;admin\&quot;,\n  \&quot;id\&quot;: \&quot;${orderId}\&quot;\n}&quot;,
   &quot;contentType&quot;: &quot;application/json&quot;,
   &quot;charset&quot;: &quot;UTF-8&quot;
 }</httpBodyContent>
@@ -43,18 +43,25 @@
    <socketTimeout>-1</socketTimeout>
    <useServiceInfoFromWsdl>true</useServiceInfoFromWsdl>
    <variables>
-      <defaultValue>'25333618-21cb-4518-919c-0c0e46d7a921'</defaultValue>
+      <defaultValue>''</defaultValue>
       <description></description>
       <id>85d5eae8-29e9-4548-b8dd-867f8404ae68</id>
       <masked>false</masked>
       <name>orderId</name>
    </variables>
    <variables>
-      <defaultValue>'2'</defaultValue>
+      <defaultValue>''</defaultValue>
       <description></description>
       <id>c93fb180-e17a-4e4e-b56d-164e113f818d</id>
       <masked>false</masked>
       <name>quantity</name>
+   </variables>
+   <variables>
+      <defaultValue>''</defaultValue>
+      <description></description>
+      <id>d4546e46-a720-4d6e-b9eb-482653f40789</id>
+      <masked>false</masked>
+      <name>userName</name>
    </variables>
    <verificationScript>import static org.assertj.core.api.Assertions.*
